requirements.txt
pymongo~=4.7.2
pydantic~=2.7.1
fastapi~=0.111.0
uvicorn~=0.29.0
starlette~=0.37.2
python-dotenv~=1.0.1
langchain~=0.1.17
loguru~=0.7.2
openai~=1.25.0